% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attractorFinding.R
\name{attractorFinding}
\alias{attractorFinding}
\title{Find all co-expressed signatures in the expression matrix}
\usage{
attractorFinding(exprs, dropout_ind, a = 5, epsilon = 1e-04,
  mi_threshold = 0.4, verbose = TRUE)
}
\arguments{
\item{exprs}{A log-transformed expression data matrix of which 
rows and columns represent genes and cells, respectively.}

\item{dropout_ind}{A matrix in which each row contains the indices
of zero entries that are suspected as dropouts.}

\item{a}{Exponent of mutual information (MI) used to create weighted
vector for metagenes, one of parameters in function 
\code{findAttractors} in package \pkg{cafr}. Default is 5.}

\item{epsilon}{Convergence threshold, one of parameters in function
\code{findAttractors} in package \pkg{cafr}. Default is 1E-4.}

\item{mi_threshold}{MI threshold (between 0 to 1) to determine the 
gene membership of signatures. Genes with greater than or equal to 
the threshold are considered to be associated with the signature. 
Default is 0.4.}

\item{verbose}{Whether to show the finding progress. Default is TRUE.}
}
\value{
An attractor list in which each element contains the 
vector of MIs between the genes and the converged metagene 
for each attractor signature.
}
\description{
This function uses heuristic exhaustive search to find
existing co-expressed signatures in the expression matrix,
with attractor finding algorithm implemented in the
R package \pkg{cafr}.
}
\examples{
data(ge_10x_sample)
data(results1)
attractors <- attractorFinding(ge = ge_10x_sample, dropout_ind = results1$dropout_ind)


}
