% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attractors}
\alias{attractors}
\title{Example output of function \code{\link{attractorFinding}}}
\format{An object of class \code{list} of length 2.}
\usage{
data(attractors)
}
\description{
An attractor list containing the two co-expressed gene signatures 
identified in \code{\link{ge_10x_sample}} by running function 
\code{\link{attractorFinding}}.
}
\details{
Each signature is a vector consisting of the normalized mutual 
information between the corresponding gene and the converged metagene.
}
\keyword{datasets}
