% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capping.R
\name{capping}
\alias{capping}
\title{Extreme values treatment}
\usage{
capping(x)
}
\arguments{
\item{x}{A vector containing numeric numbers.}
}
\value{
A vector with the same length as the input,
in which extreme values have been treated.
}
\description{
This function treats extreme values by winsorization.
}
\details{
For values that lie outside the 1.5*IQR limits, we cap 
them by replacing those observations outside the lower
limit with the value of 5th \%ile, and those that lie 
above the upper limit with the value of 95th \%ile.
}
