% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddImpute.R
\name{ddImpute}
\alias{ddImpute}
\title{Run ddImpute on an expression matrix}
\usage{
ddImpute(exprs, t = 0.2, genes = NULL, mi_threshold = 0.4, k = 10,
  verbose = TRUE)
}
\arguments{
\item{exprs}{A log-transformed gene expression data matrix,
where rows and columns correspond to genes and cells, 
respectively.}

\item{t}{Threshold (between 0 and 1) for determining 
similar cells in the step of dropout identification. 
Default is 0.2.}

\item{genes}{A vector containing gene symbols that will 
get imputed. Default is NULL, in which case all available 
genes in the matrix exprs will be imputed.}

\item{mi_threshold}{MI threshold (between 0 to 1) to 
determine the gene membership of signatures. Genes with 
greater than or equal to the threshold are considered to be 
associated with the signature. Default is 0.4.}

\item{k}{Number of neighbors to be used in the k-nearest 
neighbor regression in the step of \code{imputeByCells}.
Default is 10.}

\item{verbose}{Whether to show the progress of imputation.
Default is TRUE.}
}
\value{
The imputed version of expression data matrix.
}
\description{
ddImpute is an imputation algorithm designed for
recovering dropouts in single-cell RNA-seq data
by using information from two-dimensional relationships 
(cells and genes), as described in Zhu et al. 2019.
}
\examples{
data(ge_10x_sample)
imputed <- ddImpute(exprs = ge_10x_sample)


}
