% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDropouts.R
\name{findDropouts}
\alias{findDropouts}
\title{Identify dropout-suspected zeros from all zero entries}
\usage{
findDropouts(
  exprs,
  t = 0.2,
  genes = NULL,
  ncores = 1,
  return_J = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{exprs}{An input log-transformed data matrix. The rows and columns
correspond to genes and cells, respectively.}

\item{t}{Threshold (between 0 and 1) for determining similar cells. 
Default is 0.2.}

\item{genes}{A vector containing gene symbols that will 
get imputed. Default is NULL, in which case all available 
genes in the matrix exprs will be imputed.}

\item{ncores}{Number of cores to use. Default is 1.}

\item{return_J}{Whether to return the calculated pairwise Jaccard matrix between cells. 
Default is FALSE.}

\item{verbose}{Whether to show the progress of dropout identification.
Default is TRUE.}
}
\value{
A list with the following components:
\item{dropout_ind}{A matrix in which each row contains the indices of a dropout}
\item{J}{(optional) Calculated pairwise Jaccard Index between cells}
}
\description{
This function distinguishes zeros that are likely to be 
dropouts from those biologically true zeros based on 
pairwise Jaccard Index between cells.
}
\examples{
data(ge_10x_sample)
results1 <- findDropouts(ge_10x_sample, return_J = TRUE)
# pairwise Jaccard Index between cells
J <- results1$J
# indices of identified dropouts
dropout_ind <- results1$dropout_ind


}
