% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ge_10x_sample}
\alias{ge_10x_sample}
\title{Example single-cell RNA-seq dataset}
\format{An object of class \code{matrix} with 3774 rows and 338 columns.}
\source{
\href{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/jurkat:293t_50:50}{10X Genomics}.
}
\usage{
data(ge_10x_sample)
}
\description{
The dataset contains 3,774 genes and 338 cells, 
a subset of the 50\%:50\% Jurkat:293T cell mixture 
data downloaded from the 10X Genomics website.
}
\references{
Zheng, G.X. et al. Massively parallel 
digital transcriptional profiling of single cells. 
\emph{Nat Commun} 8, 14049 (2017). (\href{https://www.ncbi.nlm.nih.gov/pubmed/28091601}{PubMed})
}
\keyword{datasets}
