% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeByAttractors.R
\name{imputeByAttractors}
\alias{imputeByAttractors}
\title{Imputation based on co-expressed gene attractor signatures}
\usage{
imputeByAttractors(
  exprs,
  dropout_ind,
  attractor_list,
  mi_threshold = 0.4,
  verbose = TRUE
)
}
\arguments{
\item{exprs}{A log-transformed expression matrix, where
rows and columns represent genes and cells, respectively.}

\item{dropout_ind}{A matrix where each row contains the 
indices of zero entries (first column for gene name, 
second column for cell ID) that are suspected as dropouts.}

\item{attractor_list}{A list where each element contains
the identified co-expressed signature.}

\item{mi_threshold}{MI threshold (between 0 to 1) to determine the 
gene membership of signatures. Genes with greater than or equal to 
the threshold are considered to be associated with the signature. 
Default is 0.2.}
}
\value{
A list consisting of the following components:
\item{imputed}{The imputed version of the expression matrix}
\item{dropout_ind}{An updated indices matrix of dropout-suspected
zero entries, where in each contains the indices of the remaining
dropouts after this step of imputation.}
}
\description{
This function imputes dropout-suspected zeros by borrowing 
information from co-expressed genes in the identified 
co-expressed attractor signatures.
}
\details{
For each gene of interest, we use the top 20 genes involved 
in the same signature as the predictive variables to fit a 
linear regression model using cells where all those genes are
expressed. Details can be found in \url{paper}
}
