% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeByCells.R
\name{imputeByCells}
\alias{imputeByCells}
\title{Imputation based on cell-to-cell relationships}
\usage{
imputeByCells(exprs, dropout_ind, k = 10, verbose = TRUE)
}
\arguments{
\item{exprs}{A log-transformed expression matrix, where
rows and columns corresponds to genes and cells, respectively.}

\item{dropout_ind}{A matrix where each row contains the 
indices of zero entries (first column for gene name, 
second column for cell ID) that are suspected as dropouts.}

\item{k}{Number of neighbors to be used in the kNN regression.
Default is 10.}

\item{verbose}{Whether to show the progress of imputation.
Default is TRUE.}
}
\value{
An imputed version of gene expression matrix.
}
\description{
This function imputes dropout-suspected zeros by 
borrowing information from similar cells using
k-nearest neighbors (kNN) regression.
}
\examples{
# the following imputation is based on the results 
processed by function 'imputeByAttractors'
data(results2)
imputed <- imputeByCells(results2$imputed, results2$dropout_ind)


}
