% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imputed}
\alias{imputed}
\title{Example output of function \code{\link{ddImpute}}}
\format{An object of class \code{matrix} with 3774 rows and 338 columns.}
\usage{
data(imputed)
}
\description{
The imputed version of the example scRNA-seq data
\code{\link{ge_10x_example}} by ddImpute algorithm.
}
\keyword{datasets}
