% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_index.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Calculate pairwise Jaccard Index between cells}
\usage{
jaccard_index(Matrix, threshold = 0, verbose = FALSE)
}
\arguments{
\item{Matrix}{An input matrix, where rows and columns represent
genes and cells, respectively.}

\item{threshold}{The threshold for generating binary matrix.
Values greater than it are set to 1, otherwise 0. Default is 0.}

\item{verbose}{When TRUE, it will show the progress of calculation.}
}
\value{
A symmetric matrix containing pairwise Jaccard Index values
between cells.
}
\description{
This function calculates the Jaccard Index between each pair of cells
based on binarized expression matrix.
}
