% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results1}
\alias{results1}
\title{Output of function \code{\link{findDropouts}}}
\format{An object of class \code{list} of length 2.}
\usage{
data(results1)
}
\description{
The results of performing dropout identification on data 
\code{\link{ge_10x_sample}}. A list consisting of
\item{dropout_ind}{A matrix in which each row contains the indices of a dropout}
\item{J}{Calculated pairwise Jaccard Index between cells}
}
\keyword{datasets}
