% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_2DImpute.R
\name{run_2DImpute}
\alias{run_2DImpute}
\title{Run 2DImpute on an expression matrix}
\usage{
run_2DImpute(
  exprs,
  t = 0.2,
  genes = NULL,
  k = 10,
  ncores = 1,
  log_input = TRUE,
  return_J = FALSE,
  return_attractors = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{exprs}{A log-transformed gene expression data matrix,
where rows and columns correspond to genes and cells, 
respectively.}

\item{t}{Threshold (between 0 and 1) for determining 
similar cells in the step of dropout identification. 
Default is 0.2.}

\item{genes}{A vector containing gene symbols that will 
get imputed. Default is NULL, in which case all available 
genes in the matrix exprs will be imputed.}

\item{k}{Number of neighbors to be used in the k-nearest 
neighbor regression in the step of \code{\link{imputeByCells}}.
Default is 10.}

\item{ncores}{Number of cores to be used. Default is 1.}

\item{log_input}{Whether the input data matrix is log-transformed.
Default is TRUE.}

\item{return_J}{Whether to return the calculated pairwise
Jaccard matrix between cells in the step of \code{\link{findDropouts}}. 
Default is FALSE.}

\item{return_attractors}{Whether to return the co-expressed
gene attractors found by the function of \code{\link{attractorFinding}}.
Default is FALSE.}

\item{verbose}{Whether to show the progress of imputation.
Default is TRUE.}
}
\value{
A list the following components:
\item{imputed}{The imputed version of expression data matrix}
\item{attractors}{(optional) Identified co-expressed gene attractor signatures}
}
\description{
2DImpute is an imputation algorithm designed for
recovering dropouts in single-cell RNA-seq data
by using information from two-dimensional relationships 
(cells and genes), as described in Zhu et al. 2019.
}
\examples{
data(ge_10x_sample)

# Fast demonstration
res <- run_2DImpute(exprs = ge_10x_sample, genes = c('XIST', 'CD3D'), ncores = 2)
imputed_exprs <- res$imputed

# Return identified co-expressed attractor signatures
res <- run_2DImpute(exprs = ge_10x_sample, genes = c('XIST', 'CD3D'), ncores = 2, return_attractors = TRUE)
imputed_exprs <- res$imputed
attractors <- res$attractors

# Full imputation
res <- run_2DImpute(exprs = ge_10x_sample, ncores = 2)
imputed_exprs <- res$imputed

}
