% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results2}
\alias{results2}
\title{Example output of function \code{\link{imputeByAttractors}}}
\format{An object of class \code{list} of length 2.}
\usage{
data(results2)
}
\description{
The results of performing imputation using identified 
co-expressed attractors for data \code{\link{ge_10x_sample}}. 
A list consisting of 
\describe{
    \item{imputed}{The imputed version of the expression matrix}
    \item{dropout_ind}{An updated indices matrix of dropout-suspected}
}
}
\keyword{datasets}
